backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
pyjwt==2.8.0
cryptography==41.0.7
python-dotenv==1.0.0
alembic==1.13.1
python-decouple==3.8
email-validator==2.1.0

# Testing and Quality
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.5.0
httpx==0.25.2
faker==20.1.0
